Move files to representative folder
