+++ src/iac_scan_runner/enum/scan_response_type.py
