+++ src/iac_scan_runner/interface/check.py
@@ -4,7 +4,7 @@ from typing import Optional
 from pydantic import SecretStr
 
 from iac_scan_runner.check_output import CheckOutput
-from iac_scan_runner.check_target_entity_type import CheckTargetEntityType
+from iac_scan_runner.enum.check_target_entity_type import CheckTargetEntityType
 
 
 class Check(ABC):
